[ondemand] do not crash if the analysis of the default summary fails
Summary:
When analysing a procedure for a specialised summary and that procedure
has no default summary, we first compute the default summary. But if
that analysis fails, we crash. However, there could be a timeout
analysing the default summary, or any random reason why analyses fail.
We normally return `Error _` for these failures instead of crashing so
we shoud do the same in this edge case too.

Reviewed By: ngorogiannis

Differential Revision:
D63464980

Privacy Context Container: L1208441

fbshipit-source-id: facefbdc4aff122cece9b168595206f71e097a81