[inferpython] basic translation of CALL_FINALLY
Summary:
this opcode is a subroutine call which pushes
a return address on the top of the stack. This is not easy to
translate, but it seems to only appear in exception handlers.
Since we only translate the instructions reachable from
non-exceptionnal edges, we don't need to do much for now.

Reviewed By: nbenton

Differential Revision:
D63538161

Privacy Context Container: L1208441

fbshipit-source-id: e92743e31d03bff73faffded450fbd16afa08f92