[inferpython] building CFG nodes in topological order
Summary:
We change the order in which we examine each node
and built it. This weak topological order allows us to only build a node
once its direct CFG predecessors have been processed, except if the node
of interest is a loop header. This new invariant will help us implementing
an optimization about node parameters that is coming in a next diff.

Reviewed By: skcho

Differential Revision:
D63633175

Privacy Context Container: L1208441

fbshipit-source-id: d7a8fb106a03c9bef239a4c827ae345cf7993b71