[uninit method] Introduce a new checker
Summary:
This diff introduces a new checker that detects uninitialized method calls in Hack.

Problem: The root cause of the issue is the same with that of the uninitialized const issue; we can invoke a static function on an interface or an abstract class, and sometimes an uninitialized/undefined static method is invoked. While some simple cases can be caught by the type checker, when a type variable is used for the classname, it misses the issue, e.g., `$c::foo()`.

How to analyze: The checker simply lookups whether the callee is in Infer's type environment (`tenv`) or not.

- If a struct type info for the classname is not found, which means the class/interface was not fully captured. We do not report an issue in this case.
- If a dummy struct type info is found, we ignore the method call.
- If a struct type info is found, but the method being called is not in the list of the defined methods, we report an issue.

Note that current checking is done intra-procedurally. We will extend it inter-procedural in the following diffs.

Reviewed By: davidpichardie

Differential Revision: D62193369

fbshipit-source-id: 76ed4b4f0cab2daab7a40c9fa138a6494e999d0b