[inferhack] avoid introducing some SSA parameters
Summary:
Currently, each time the symbolic execution
reaches a new node, it cleans the current symbolic operand stack
and replaces it with a stack of the same size containing only
fresh SSA variables. The corresponding node is  given these
variables as `node parameters`. This is correct but it sometimes
lead to overcomplicated IR and we miss some opportunities of
IR simplification. It leads to LOAD_ERROR_METHOD errors.

Now, before building a new node, we examine all the incoming symbolic stacks and detect portion of stack that are equals for all predecessors. We only care about bottom of stack because this the pattern the compiler will generate often.

Reviewed By: nbenton

Differential Revision:
D63683138

Privacy Context Container: L1208441

fbshipit-source-id: 703a96fd39a0980fb057132df3af906e41ca28de