Use only one order for abstract values
Summary:
In `PulseFormula`, we used to have a custom `VarMap` with a different
order from the `Var.Map` used in other places
(`Var`=`PulseAbstractValue`). This is not necessary, and potentially
confusing.

All we need is to make sure that both `var_eqs` and `linear_eqs`, when
used as substitutions, produce as many restricted variables as
possible, so that we derive as many facts for the tableau as possible.
This can be done based on the natural order, which says that restricted
variables are smaller than unrestricted ones. It's only a matter of
using max/min in the right places.

Reviewed By: jvillard

Differential Revision: D63829660

fbshipit-source-id: 9adfcc945e4d1b3c5ef371980c303d3f987a4ff2