[erlang] Fix encoding of 'φ is true'
Summary:
We often encoded 'φ is true' by saying 'φ=v and v>0' for some fresh
variable v. The main reason is code organization: some part of the code
generates φ but returns only v to the caller, which adds the v>0
condition. But `PulseFormula` rightly does not infer that φ should be
true. Indeed the encoding of truth values should be complimentary; that
is, if 'true' is '>0' then 'false' should be '<=0', which is weird.

The convention that works better is 'true' is '!=0' and 'false is '=0'.
And with this convention the prover does what we expect.

The outcome: a lot of states that were infeasible are now detected as
such.

Reviewed By: thizanne

Differential Revision: D63838812

fbshipit-source-id: b45c82ec0ceb8d4f484006bc01ae9287cb4b6f2f