[clang] Add function prototype to Objective-C blocks
Summary:
We were passing `attr_info` too much incorrectly. Now only passing it in the few places where it's needed to go from AttributedType to BlockPointerType.

Now also passing ?from_block to only translate the function prototype for blocks and not for other function pointers, since that's not required at the moment.

Translating the function prototype of blocks including its annotations helped handle correctly the example from the test in the parameter not null checked checker.

Reviewed By: ngorogiannis

Differential Revision: D64107506

fbshipit-source-id: 6980396bb01db7ac6bc70fadcae9de161e9fbe1e