[clang] Add is_implicit field to Lfield expressions that correspond to ivars
Summary: Use the newly defined `is_implicit` field when `self` is implicit in ivar accesses. We know that this is the case when the source range from the node `ObjCIvarRefExpr` includes the same column at the start and at the end. If `self` is explicit, the columns are different. I checked this in the bdump. Unfortunately there are no other nodes or fields that are different between the two cases.

Reviewed By: geralt-encore

Differential Revision: D64404896

fbshipit-source-id: b1edcfb6d8dbb524e6071b71cdec5e45fd822be3