[self-in-block] Improve autofix for when self is implicit in ivars
Summary:
Use the `is_implicit` field of the object in `Lfield` to differentiate the autofixes between when self is explicitly there vs. when it's implicit.

For instance we have the following situation

```
n$225=*&self:SelfInBlockTest* [line 370, column 21];
n$226=*n$225._name:NSString* [line 370, column 21];
```
We add `n$225` to Vars in the first line, and in the second line, we retrieve it again, and add the `is_implicit` flag if needed.

In the implicit case, the autofix is

original: ""
replacement: "strongSelf->"

which should deliver the expected result.

Reviewed By: geralt-encore

Differential Revision: D64405099

fbshipit-source-id: 86d8dab97390eb08a6fdf90f92388a7702a81d73