[ondemand] more deterministic mutual recursion
Summary:
The main source of non-determinism in infer comes from mutual recursion: the analysis will start from somewhere in the cycle and produce different results depending on where it started. This is (worse) because we stop at the first recursive call and return an empty summary for that one, which makes the place where we start in the cycle all the more important.

Restart the analysis from a predictable place (the smallest procedure) when a mutual recursion cycle is detected so that we always(?) get the same result. If the recursive structure is complex it may restart a bunch of times but it should always eventually terminate since we must pick a smaller procedure each time (assuming analysing a procedure will always generate the same recursive ondemand requests every time, *cough* that might be a problem...).

Reviewed By: skcho

Differential Revision:
D63633971

Privacy Context Container: L1208441

fbshipit-source-id: b4cae047d0275907bcdf0ac2c626828bd9aa80dc