[param-not-null-checked] Only add autofix if the return type of the block is void
Summary:
Only add autofix if the return type of the block is void. The reason for this is that the autofix we use is a macro, and replacing the block call with the macro doesn't compile unless the return type is void.

We were not translating the return type of the block in all cases, a small change was needed to translate it in the case of the example in the test.

Reviewed By: geralt-encore

Differential Revision: D64964204

fbshipit-source-id: d09598b18bff66f45552e00d15d9f496cee42208