[textual] normalize textual code generated during closure transformation
Summary: Textual provides a closure construct to ease frontend work. Each closure is then translated into a OO encoding. But this transformation was not fully tested and I realize it was not fully correct. It needs to use a specific form because it comes after some transformations in the pipeline.

Reviewed By: skcho

Differential Revision:
D64969396

Privacy Context Container: L1208441

fbshipit-source-id: ca24a7e0393361c7e49dfa26ed16b54e56113c5e