[nstring-internal-ptr] Adding a new issue type NSSTRING_INTERNAL_PTR_CAPTURED_IN_BLOCK
Summary:
The issue is that 	when the result of a call to the method `cStringUsingEncoding` of `NSString` or its property `UTF8String`
is captured in an escaping block without copying it first, this can cause a crash.

We adapt the checker for finding CXX String Captured In Block to deal with this case too as both cases are very similar.

First,  we add the context info to the captured variables in the preanalysis, and then flag it in the checker.

Reviewed By: geralt-encore

Differential Revision: D65070300

fbshipit-source-id: 87dde7d90e8e22f69ad23afe78cdea90c50f5c23