[inferpython][3.10] opcode GEN_START
Summary:
this is a coroutine opcode that seems to be inserted at the beginning
of function using this feature. As explained in a comment, the current implementation
is not completely aligned with the official spec but we do not model coroutine very precisely
anyway for now.

Reviewed By: skcho

Differential Revision:
D65015190

Privacy Context Container: L1208441

fbshipit-source-id: cfd243beeadd4768d1e17f1982155cad02da01ab