[pulse] Address infer_skip as unknown call
Summary:
The builtin `infer_skip` is introduced when the frontend could not properly translate the code to
SIL.  Previously, this was modelled as `Basic.skip`, but which did not reflect the unknown sematnics
of the code.  This diff changes it to be modelled as `Basic.unknown_call` instead.

```
n$6 = infer_skip(...) // This is modelled as an unknown call.
*x = n$6 // On the following store statement, we propagate the unknown effect.
```

Reviewed By: jvillard

Differential Revision: D65210357

fbshipit-source-id: 08d817c3f360c58d8855eb34ed8306850f6b7bf0