[self-in-block] Refactor of stringSelfNotChecked code to allow for dedup
Summary:
We move reporting of stringSelfNotChecked issues to the end as with the other issue types to be able to keep a set of reported variables and report only once.

This also simplifies the algorithm and changes where we report. We were avoiding reporting in places that are ok with nil, such as method calls. However, this is a code pattern, and they should add the check for nil as a good practice, because other code may be added later and the check will be forgotten.

Reviewed By: skcho

Differential Revision: D65218571

fbshipit-source-id: e190e98facbefa245bd48320e4f64de7eec2783b