[hack] fix arity of trait calls when resolving optional arguments
Summary:
Trait methods with optional arguments generate wrapper methods that call
the "base" method with no optional arguments. However they do this via a
virtual call (which is not ideal and should be fixed) and the arity of
that call is not set properly and fails to account for the extra
`self` argument that is added later by the call resolution logic in
pulse. As a result we get a spurious recursive call every time such a
method is defined (it doesn't even need to be called!), which turns into
an unknown call (and a spurious MUTUAL_RECURSION warning). This isn't a
big problem for the analysis because many of these methods will end up
being specialised and in that case the inner call gets properly
resolved.

This fixes the arity of the procedure called by detecting if we are in a
virtual call *and* in a trait and if so adjusting the arity. This should
correspond mostly to this case, but it's a hack and could mess up other
things.

Reviewed By: davidpichardie

Differential Revision: D65613649

fbshipit-source-id: 148f68631e7b96cc7bd9c23cdcf5e59bd9180c87