[clang] Fixing the perf regression in the captured context preanalyses and reenable them
Summary:
So we have two analyses with their domains. We want to process those domains to add some data to the attributes of some blocks. I believe that the way it was done before was inefficient. Now we pre-process the domains to only have a simple data structure with the data that we actually need, and we preprocess the procedures to keep only the blocks. Then we process one domain after the other. This should be ok now.

After enabling the preanalyses again, we are able to report BLOCK_PARAMETER_NOT_NULL_CHECKED on captured variables again, as well as CXX_STRING_CAPTURED_IN_BLOCK and NSSTRING_INTERNAL_PTR_CAPTURED_IN_BLOCK.

Reviewed By: ngorogiannis

Differential Revision: D65942978

fbshipit-source-id: 3424fc73841fc4a3dcf8ff85eaa5818e440c69c8