[self-in-block] Do not report on instance method calls
Summary: We want to stop reporting when strongSelf is the receiver of instance method calls because that's not a crash. Previously we were checking the virtual call flags but that was not enough for checking getters and setters of properties, so we are now also checking the instance property of the callee method.

Reviewed By: ngorogiannis

Differential Revision: D66101550

fbshipit-source-id: f11e248845ef698a95695a73140dc36608fbc919