[hack] Modify function reference arg only when needed
Summary:
When `$f(...)` is invoked, we modify the receiver as,

```
lateBinding($f, ...)
===>
lateBinding($f.this, ...)
```

to pass a proper function reference to `lateBinding`.  However, this
modification was applied unconditionally, e.g.

```
hhbc_new_vec($f, ...)
===>
hhbc_new_vec($f.this, ...)
```

which introduced a wrong abstract semantics.

This diff changed it to modify the receiver only when the method name is `lateBinding`.

Reviewed By: geralt-encore

Differential Revision:
D66235651

Privacy Context Container: L1208441

fbshipit-source-id: 90ccfd77be393af2f1226e1da76246ab22541fc5