[dispatch-once-static-init] Improve domain with traces
Summary:
We keep a list of traces as the main element of the domain. We only add an element to the domain when there is a call to `dispatch_once` or when the summary is non-empty.
Once we report an issue (only in a static initializer), we then don't keep track of the calls in the summary anymore.

Reviewed By: jvillard

Differential Revision: D66537767

fbshipit-source-id: 68ac7560cb0507ea05d99763fa2b63169fac666e