[inferpython] generate a default type for each module
Summary:
any function takes as argument a paremeter 'globals'
and every call to a function or an imported module will go through
this dictionnary. Currently, this oblige us to ask for a
specialization key that is given by an explicit call site (for example
from the module body). We can avoid this need if we annotate the
parameter with a type, as we do in Hack.

This "default" type is guessed by looking at the module body. This
 is highly dependent on the Python bytecode compiler. I plan to add
a security mechansim that will check if the typ is valid or not, by such
a safety belt is not something urgent.

Reviewed By: ngorogiannis

Differential Revision:
D66600358

Privacy Context Container: L1208441

fbshipit-source-id: f22a47ec174b39abc7003407a5b4f68167ac8894