[pulse] minor refactor in PulseFormula
Summary:
We will use this in the next diff to manipulate term equalities and
their occurrences (and atoms and their occurrences) more directly.

Also record that `x` occurs in the range of `term_eqs` when `l=x` is
added to `term_eqs` and `l` is linear. Somehow we didn't need this
before but this is going to be needed too.

Reviewed By: ngorogiannis

Differential Revision: D66710758

fbshipit-source-id: 81300fa305d25ccb37e93013e86759a97468d4a3