[inferpython] restore the previous chaining of globals value
Summary:
In a previous diff, I changed the chaining of globals parameter because I was facing strange alias specialization behaviors.
But the new module object encoding seems to have make this problem disappear. I'm switching back to the old convention and it removes a FN.

The other chaining is semantically unsound because it gives the wrong globals to a call `foo()` if `foo` comes from a `from .. import foo`.

Reviewed By: jvillard

Differential Revision:
D66758255

Privacy Context Container: L1208441

fbshipit-source-id: b036db24947fa9df16fc47fcf2198fee7f9d8554