[clang] detect objc getter/setter calls
Summary:
clang's treatment of getters and setters in ObjC is pretty goofy:
accessing an objc "property" (=field) is always translated into a getter
or setter call, even from within the definition of custom getters and
setters! Hilarious.

We want to know when we are calling a getter or setter so that we at
least don't report infinite recursion on these calls.

Clang knows when we're making a getter/setter call, but the information
is placed awkwardly in the AST:
```
  `-PseudoObjectExpr
   |-ObjCPropertyRefExpr Messaging=Getter // WE WANT TO KNOW THIS
   | `-...
   |-OpaqueValueExpr
   | `-...
   `-ObjCMessageExpr // BUT WE WANT TO KNOW THIS HERE
    `-...
```

We need to propagate the information from the `ObjCPropertyRefExpr` to its
sibling `ObjCMessageExpr` within `PseudoObjectExpr`.

Reviewed By: skcho

Differential Revision:
D66885510

Privacy Context Container: L1208441

fbshipit-source-id: 73844626f0356f56f4b862b5985f9e86cb929ead