[pulse][join] set over-approx state to Bottom at the start of each instruction
Summary:
This is needed for the soundness of executing over-approximate states:
otherwise we copy the pre over-approx state unchanged to the post of
each disjunct, and join it (the pre over-approx state) to the post
over-approx state computed by executing the pre over-approx state on its
own. This was wrong.

The solution is simple: prevent disjuncts from seeing a
potentially-outdated over-approx state by setting it to Bottom, knowing
the true over-approx state will get executed and joined separately.

Reviewed By: skcho

Differential Revision: D66974651

fbshipit-source-id: 217e5466faa6f375a0603618d40f6703a30fe625