[inferpython][specialization types] adding a new type kind for import_from
Summary: we want to avoid a need for specialization when a function use an other function `fun` obtained by a `from .. import .. as fun`. This is not trivial because the closure type of `fun` is not computable without searching inside the Tenv, while we generate specialization types at capture time. We introduce a *delayed* name that will be resolve later at analysis time (see next diff).

Reviewed By: skcho

Differential Revision:
D67139193

Privacy Context Container: L1208441

fbshipit-source-id: fe46377766fe2a4fdb033460cd747b42606e65e1