[multicore] avoid racing on chdir
Summary:
In multicore we can't allow the workers to change the working directory concurrently in `Utils.do_in_dir` as they will eventually execute their functions in the wrong directory.

This diff introduces a mutex for exclusion, and removes redundant uses of the function which can crash due to recursively locking the mutex.

Reviewed By: jvillard

Differential Revision:
D67195918

Privacy Context Container: L1208441

fbshipit-source-id: 6f7c3518fbc9b638e8a5e71fa15bf4ce621ffe22