[inferpython] support for import-package
Summary: import-package and from-package-import-attr behave quite differently. In the first case, the global scope is given an attribute "root" (where "root" is the name of the root of the package) which contain a package object. In the latter case, this intermediate object is not built.

Reviewed By: martintrojer

Differential Revision:
D67555258

Privacy Context Container: L1208441

fbshipit-source-id: 875da04c37e662a71bf2462d6bb5fbb504fed47d