[multicore] test-bed analysis scheduler
Summary:
A test-bed analysis scheduler that works like the file scheduler, but uses domains for analysis workers.
It's not meant for real use as it misses a number of features supported by the multiprocess analysis.
The core function is to use it with the Infer test suite, find crashes/races and generate fixes.

For example
```
$ INFER_ARGS=--multicore make direct_python_pulse_test
```
works
where
```
$ INFER_ARGS=--multicore make direct_cpp_pulse_test
```
crashes (non-deterministically)

Reviewed By: davidpichardie

Differential Revision:
D67012774

Privacy Context Container: L1208441

fbshipit-source-id: 0c61610e22449a418a31107b5701e20337e02bd7