[multicore] slight refactor of analysis global state
Summary:
Domain local storage is always declared with an initialisation function, which is called when accessing said storage for the first time from a new domain.

AnalysisGlobalStorage allows specifying an initialisation function as well, so now there is a possibility to create domain-local storage initialised by one function when spawning a domain, and initialised with another when recursing into the analysis of a new procedure. This was not possible before the introduction of DLS and is not reflected into the type system (the type `'a DLS.t` doesn't reflect the identity of the initialisation function).

This diff plugs this hole by creating and returning the DLS key inside `register_dls` function. Unfortunately we can't do the same for `register_dls_with_proc_desc_and_tenv` since the type of the initialiser function differs (`Procdesc.t -> Tenv.t -> 'a` vs `unit -> 'a`).

Reviewed By: jvillard

Differential Revision:
D67974956

Privacy Context Container: L1208441

fbshipit-source-id: 9653f1fceb8aba6037be77afcb492fb4501f40b6