[multicore][summary] do not cache incomplete summaries
Summary:
Caching an empty summary when the summary is first created creates the potential for races since the cache is shared.
This diff stops caching empty summaries and changes the logic of storing summaries so that mutations happen before caching.
Summaries remain mutable and specialisation, in particular, may still create problems, but this is for the future.

Reviewed By: jvillard

Differential Revision:
D68198006

Privacy Context Container: L1208441

fbshipit-source-id: 297e9e5941201810efb53b477ac9c169ce86f629