[multicore][tasks] polymorphic child_id type
Summary: TaskGenerator is a bit too specialised for multiprocessing: it expects a Unix pid passed to the `next` function. This will not make sense in a domain-based implementation (all the pids will be the same), so we make that type polymorphic. ProcessPool instantiates that type to a `Pid.t`.

Reviewed By: jvillard

Differential Revision:
D68446335

Privacy Context Container: L1208441

fbshipit-source-id: 9977ffd4265da122f6efb85dd9ae0b9b6ed3d22a