[pulse] take globals into account
Summary:
When checking if a MUTUAL_RECURSION is an INFINITE_RECURSION, we check
if the values of the parameters of the recursive call are the same as
the values we started with. But parameters are not the only inputs to a
procedure: sometimes they also manipulate globals, which should be
treated the same way as formals.

Reviewed By: davidpichardie

Differential Revision: D68505069

fbshipit-source-id: a4c7c5eb5765c22b23c0a9bdb157271b3ebd1817