[inferpython][textual] adding an annotation in front of each toplevel closure
Summary:
each `def foo(...):` statement leads to the generation of a closure expression.
We want to influence the name of the type that will be generated for each such closure. This
diff initiates this feature by adding an attribute. We do not use it yet.

Reviewed By: Gabisampaio

Differential Revision:
D68491365

Privacy Context Container: L1208441

fbshipit-source-id: ed963d58267b39139f6eb3882d1975999b45a733