[inferpython] catch reserved callees in py_call builtin
Summary:
we want to model some function call, even in the situation where the implementation of the callee has been captured.

Imagine we want to catch the specific call to `wait` defined in module `dir1.dir3.testmod`.  (see the added test in `async_import_with_package.py`)

At call site, we will just see something like:
```
      n7:*PyGlobals::async_import_with_package = load &globals
      n0 = $builtins.py_make_none()
      n3 = $builtins.py_load_global("wait", n7)
      n4 = $builtins.py_load_global("asyncio", n7)
      n5 = $builtins.py_call_method("sleep", n4, n0)
      n6 = $builtins.py_call(n3, n0, n5)
```
The last instruction is the call we want to catch.

During the analysis of `n3=$builtins.py_load_global("wait", n7)` we use a type propagation to attach the static type we find in
the file `dir1.dir3.testmod`, for attribute `wait`. We can later scan this name, in model `call` to recognize the name we are looking for.

Reviewed By: Gabisampaio

Differential Revision:
D67760234

Privacy Context Container: L1208441

fbshipit-source-id: 1188dc8d8b120477300934284d7daf0110433a3d