[inferpython][await iterables] builtin py_call_function_ex
Summary:
We currently have a bunch of FP of the form
```
asyncio.gather(*a_tuple)
```
Such a call uses the special builtin `py_call_function_ex` that takes a tuple of args. We install
the mechanism that will allow us to catch the name `('asyncio', 'gather')` and will install a model
for this python function in a next diff. Note that in this situation, the implementation is not captured.

Reviewed By: Gabisampaio

Differential Revision:
D67788681

Privacy Context Container: L1208441

fbshipit-source-id: c60b6d1d5560d5052c5b1cfd4ed8170c211119de