[topl] Added ability to negate regexps
Summary:
Regexps are used in two situations, to match procedure names and to
match types. Matching a regexp R against a procedure name P is done the
same way as matching against a string P. For the negated regexp !R, we have
`(!R)~P` when `!(R~P)`.

Matching against types is slightly more involved. Write `R~~T` for "regexp
R matches type T", as opposed to R~S which means "regexp R matches
string S". Then `R~~T` when `R~name(S)` and `T<:S` for some type S. And
`(!R)~~T` when `!(R~~T)`. (For subtyping, arrays are considered covariant,
but this is unrelated to this diff -- it's just a pragmatic
simplification preexisting in the code.)

Reviewed By: hajduakos

Differential Revision: D68781862

fbshipit-source-id: c6700990906c339e5e4f57d295884871ec32b6c7