[pulse] ensure we log why formulas become UNSAT
Summary:
To be really sure we never forget, change the type of `SatUnsat.t` to
include an explanation for `Unsat`.

We may want to extend this to record some more info, eg for types we get
"0≠1" as explanation instead of "x doesn't have type T" because of how
typing judgements are encoded.

Reviewed By: geralt-encore

Differential Revision: D69054477

fbshipit-source-id: a5f9f8f5fb7c5d84cf972d3051d8d133c8d0e656