[swift] Llair to Textual: translating formals and locals from functions
Summary: We aim to translate llair functions, and we'll do it in parts. Here we do formals and locals (except for the types). We declare a temporary data structure `partial_proc_desc` to keep the translated parts and make sure that we have the required types in Textual correctly. So we can keep adding fields until we have complete proc_desc, similarly for proc_decl.

Reviewed By: davidpichardie

Differential Revision: D69127884

fbshipit-source-id: 7c976039b61ed3045771af5d15933228467dba7b