[inferpython] do not track unwaited awaitable inside generators
Summary:
generators will be quite challenging to track precisely because of
the yield instruction that stop the current iterator until next iteration.

In the context of the UnwaitedAwaitable checker, we just want to forget every unwaited awaitables that are returned by such a function.

A Python expression like
`( asyncio.sleep(i) for i in l )`
is compiled into a call to
```
{
  #b0:
      n2 = globals
      n1 = locals
      n0 = $builtins.py_make_none()
      _ = $builtins.py_gen_start_generator()
      n3 = $builtins.py_load_fast(".0", n1)
      jmp b1

  #b1:
      n4 = $builtins.py_next_iter(n3)
      n5 = $builtins.py_has_next_iter(n3)
      if n5 then jmp b2 else jmp b3

  #b2:
      _ = $builtins.py_store_fast("i", n1, n4)
      n6 = $builtins.py_load_global("asyncio", n2)
      n7 = $builtins.py_load_fast("i", n1)
      n8 = $builtins.py_call_method("sleep", n6, n0, n7)
      n9 = $builtins.py_yield(n8)
      jmp b1

  #b3:
      _ = $builtins.py_nullify_locals(n1, "i")
      ret n0

}
```

Hence we have to implement a model for `py_yield` here.

Reviewed By: ngorogiannis

Differential Revision:
D69176729

Privacy Context Container: L1208441

fbshipit-source-id: 44024613ae85020861f6b4d6d0b6e4a35882e6c5