[multicore] safe lazy evaluation
Summary:
The previous attempt to fix concurrent lazy evaluation is insufficient: the code either needs to indefinitely retry, or use another mechanism. Putting a ceiling in the number of retries makes this algorithm flaky, and that's the whole point of the method of avoiding race conditions.

This diff introduces a wrapper for lazy expressions that protects forcing by way of a mutex. Some care is taken to avoid allocating a mutex if creating a lazy expression out of a concrete value. Also, some care is needed when wrapper values need to be marshalled, since mutexes will make this fail.

Reviewed By: davidpichardie

Differential Revision:
D68204198

Privacy Context Container: L1208441

fbshipit-source-id: 5c3468b349b96283f5fb5c017ae70af23f0e27e3