[serialization] remove unused functionality
Summary: The only use of the `Serialization` module is in `Tenv`. To use this module effectively, one needs to update manually some magic numbers every time the serialization format changes. This is rarely, if ever, done. In addition, this module is not used when serialising into the database. To simplify things we remove this module and fall back to plain marshalling.

Reviewed By: dulmarod

Differential Revision:
D69182709

Privacy Context Container: L1208441

fbshipit-source-id: bddfa50cb226468d3b7faf1df037fb44db392670