[swift] Llair to Textual: translate llair types
Summary: We translate the types, llair types are richer, we may need to add more information in the future. They also include the whole struct information, which we store in a separate data structure, so we'll add each struct to our struct declarations as we encounter it, in a follow up diff.

Reviewed By: ngorogiannis

Differential Revision: D69253392

fbshipit-source-id: 0f90469683c264a76ab71025e02cbb4860fe4bfd