[infer]Llair to Textual: translate structs
Summary:
We keep a global map of structs that we fill when translating types. At the end of the translation we use this map to update the textual module with the structs.

Only named ones so far, we'll do llair's tuples or annonymous structs in a follow up diffs.

In llair, we only have access to field offsets rather than names, so we create internal names for the fields to add them to textual. We hope to get the field names from llvm later on.

Reviewed By: davidpichardie

Differential Revision: D69403904

fbshipit-source-id: 4743452b1c6489111f1a680e96cf640984e1db34