[multicore] configurable caching policy for attributes
Summary:
Caches are treated differently in process mode and in multicore mode:
In process mode, the caches are cleaned after every target, to avoid taking too much memory.
In multicore mode, the caches are shared, and therefore cleaning them reduces the benefit.

This diff makes the procedure attributes cache behave as usual in process mode, but in a bounded LRU mode in multicore mode.
We also add code to log hit rate statistics in a uniform manner for caches.
The default maximum size of the LRU cache in multicore mode (here, 500) was computed as the (rough) minimum that would guarantee the same hit rate in multicore mode as the (aggregate) hit rate in process mode, for selected code bases.

Reviewed By: davidpichardie

Differential Revision:
D69409519

Privacy Context Container: L1208441

fbshipit-source-id: 2d17f7386098ae33183f402f353cfea7ca9b520b