[summary] make into a 2-layer cache
Summary:
The summary cache is currently a map from `(procname, analysis_request) -> summary`. This requires a linear scan of the whole cache to remove entries related to a `procname`.

This diff turns the cache into a 2-layer map: `procname -> analysis_request -> summary` and removes the need for filter/linear scanning.

NB this is not totally race-free, because the critical section protects the first level only. The next diff addresses this.

Reviewed By: davidpichardie

Differential Revision:
D69458867

Privacy Context Container: L1208441

fbshipit-source-id: 22260fe1491d30699e014226ca823b3376c75bf2