[summary][multicore] configurable caching policy
Summary:
Turn the summary cache into an LRU cache in multicore mode (does not affect process mode).
One particularity here is that to retain race freedom in a 2-layer cache, we lock the table while updating both layers, not just the first one.
Lookups are not protected in the same way, but since the 2nd layer is an immutable map, races here will just mean work duplication which is already possible (depending on scheduler).

Reviewed By: davidpichardie

Differential Revision:
D69460473

Privacy Context Container: L1208441

fbshipit-source-id: 1f6c43d6d5ffa0fcfc039fe61f216fc9d33e13e5