[qualcppname] fewer allocations
Summary:
Multicore performance is particularly sensitive to excessive allocations.
This diff removes a lot of allocations from `QualifiedCppName.compare_name` and callers, by using `String.index '<'` and a substring operation instead of `String.split ~sep:'<'` to remove template arguments. The result is zero allocations in the case of a non-templated argument.

In a test target this leads to about 400 fewer minor collections (over a baseline of 6200). memtrace (after backporting to 4.14) shows a much bigger reduction of total allocations (from 24B to 20B) and a complete disappearance of `String.split` from the top 10 allocators (it was 1st).

Reviewed By: davidpichardie

Differential Revision:
D69594093

Privacy Context Container: L1208441

fbshipit-source-id: b679dfe7feff75f24928e4a5887f0b64c4d7477e