[swift] Llair to Textual: translate and collect all the nodes in a proc
Summary:
In llair, there is only a reference to the entry block, so when we translate blocks to Textual nodes, we also need to collect the nodes to be able to translate them and add them to the proc_descs.

So we translate nodes, starting from the entry node, and following jumps, translating nodes as we go and collecting the successors. We keep a set of node names of seen nodes to make sure the recursion stops in the presence of loops.

Reviewed By: davidpichardie

Differential Revision: D69785730

fbshipit-source-id: 771a6dcfa3b12042f8abc4515b7669e8830515f9