[sledge] fix C bindings causing segfaults
Summary:
`to_val()` must be used to go from C++ LLVM values to OCaml ones as it
flips a bit to tell OCaml not to assume ownership of it. This caused
segfaults.

The rest are mostly style convention obeyed in the rest of the file I
think, but also cannot hurt.

Reviewed By: ngorogiannis

Differential Revision: D70090791

fbshipit-source-id: 5d64a7bb01cc688cfad5e9df9933d0b7db95161f