[swift] Llair to Textual: translate alloc
Summary:
Looking at `sledge/test/analyze/free_free.c`, a code like

```
  int* p = (int*)malloc(sizeof(int));
```

gets translated to llair as

```
&main () { #1 ; test/analyze/free_free.c:9:
    %1!3 := alloc [((i64)(s32) 1) x 2];	; test/analyze/free_free.c:11:8:
    %2!5 := builtin malloc(4);	; test/analyze/free_free.c:11:18:
    store 8 %1!3 %2!5;	; test/analyze/free_free.c:11:8:
}
```

but with the clang frontend we do

```
n$4=_fun_malloc(sizeof(t=int;nbytes=4;nullable=false):int) [line 11, column 18];
*&p:int*=n$4 [line 11, column 3];
```

so I think we probably don't need this in Infer, and it's definitely different to our standard `alloc` builtin, so I use a different name. We may want to translate as returning a fresh id to work with. But we need to see more examples first.

Reviewed By: ngorogiannis

Differential Revision: D70103763

fbshipit-source-id: 856b9d759522dadb69111f942545fa6f8b3181f5